+++ src/io/uci.cpp
@@ -113,7 +113,7 @@ void uci_parse_go(Board &board, std::istringstream is, SearchInfo &info) {
                          static_cast<unsigned long long int>(inc);
     }
 
-    info.depth = (depth == -1) ? MAX_DEPTH : static_cast<uint8_t>(depth);
+    info.depth = (depth == -1) ? MAX_DEPTH : depth;
 
     printf("info string time:%d start:%llu stop:%llu depth:%d timeset:%d",
            time,info.start_time,info.stop_time,info.depth,info.time_set);
